requirements.txt
fastapi
uvicorn
python-dotenv
chromadb
pypdf
pandas
python-multipart
requests
pytesseract
Pillow